Added routers, navigation, basic implementation of     done assets so far #14
