+++ crypto-gaze/app/react-ui/package.json
@@ -15,7 +15,10 @@
     "axios": "^0.24.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
+    "react-native": "^0.71.5",
+    "react-native-web": "^0.19.1",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "^5.0.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
