+++ crypto-gaze/app/react-ui/src/App.js
@@ -1,4 +1,5 @@
 import { useState } from 'react';
+import React from 'react';
 import { MyApp } from "./MyApp";
 function App() {
   const [inputVal, setInputVal] = useState('');
