+++ crypto-gaze/app/react-ui/src/Layout/Sidebar.jsx
@@ -1,0 +1,28 @@
+import { CardContent } from "@mui/material";
+import { Card, SideNavigation, SideNavigationItem } from "@ui5/webcomponents-react";
+import React from "react";
+import { Text } from "@ui5/webcomponents-react";
+import { useNavigate } from "react-router-dom";
+
+
+
+export function Sidebar() {
+
+    const navigate = useNavigate();
+
+    const handleClick = (event) => {
+        navigate('/'+event.target.id);
+    };
+
+
+    return (
+        <>
+            <SideNavigation>
+                <SideNavigationItem text="Home" onClick={handleClick} id="home"/>
+                <SideNavigationItem text="Legacy Content" onClick={handleClick} id="legacy_content" />
+            </SideNavigation>
+        </>
+    );
+};
+
+export default Sidebar
